refactor: 화면과 api 별도로 분리
